content-engine/requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
anthropic==0.23.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0
schedule==1.2.1
